{"schema":"decodex/commit/1","summary":"Move downward registration helpers","authority":"manual"}
