+++ packages/rsnap-overlay/src/scroll_capture.rs
@@ -4,6 +4,7 @@ mod downward_candidates;
 mod downward_committed_recovery;
 mod downward_followup;
 mod downward_recovery;
+mod downward_registration;
 mod downward_resolution;
 mod fingerprint;
 mod image_stack;
